Накостылил фабрику, накостылил стейты, накостылил хп бар у врага и отображение состояния юнитов

+++ Assets/Scripts/BuildingProfile.cs
@@ -9,11 +9,11 @@ public class BuildingProfile : ScriptableObject
     [SerializeField] private GameObject _buildingPrefab;
     [SerializeField] private string _name;
     [SerializeField] private Sprite _sprite;
-    [SerializeField] private float _price;
+    [SerializeField] private int _price;
 
     public string Name => _name;
     public Sprite Sprite => _sprite;
-    public float Price => _price;
+    public int Price => _price;
     public GameObject BuildingPrefab => _buildingPrefab;
 
     public List<ProductionElement> PossibleProduction;
